refactor(ios): extract MockServer and TestHelpers for XCUITest reuse
Move MockIssueCTLServer, FailureBox, and shared helper functions into
Helpers/ subdirectory so upcoming test files can import them without
duplication. Extend mock server with new endpoints (state, comments,
priority, labels, PR detail, end session, repo CRUD, draft PATCH) and
mutable state needed for expanded test coverage.
